File Structure changed: - Moved functions to own packages.
