+++ app/src/main/java/com/exercise/matipv2/MainActivity.kt
@@ -3,44 +3,14 @@ package com.exercise.matipv2
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Switch
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
+import com.exercise.matipv2.layout.MainLayout
 import com.exercise.matipv2.ui.theme.MaTIPv2Theme
-import java.text.NumberFormat
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -51,142 +21,17 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    MatipLayout()
+                    MainLayout()
                 }
             }
         }
     }
 }
 
-@Composable
-fun MatipLayout() {
-    /* hoisting the state for EditNumber func */
-    var amountInput by remember { mutableStateOf("") }
-    val amount = amountInput.toDoubleOrNull() ?: 0.0
-    var tipPercentInput by remember { mutableStateOf("") }
-    val tipPercent = tipPercentInput.toDoubleOrNull() ?: 0.0
-    var roundUp by remember { mutableStateOf(false) }
-    val tip = calculateTip(amount, tipPercent, roundUp)
-
-    Column(
-        modifier = Modifier
-            .padding(40.dp)
-            .verticalScroll(rememberScrollState()),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Text(
-            modifier = Modifier
-                .padding(bottom = 16.dp)
-                .align(alignment = Alignment.Start),
-            text = stringResource(R.string.calculate_tip),
-            fontWeight = FontWeight.Bold,
-        )
-        // Function to edit Bill Amount
-        EditNumber(
-            modifier = Modifier
-                .padding(bottom = 15.dp)
-                .align(alignment = Alignment.Start),
-            label = R.string.bill_amount,
-            value = amountInput,
-            onValueChange = { amountInput = it },
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Next
-            ),
-            leadingIcon = R.drawable.attach_money
-
-        )
-        // function to edit Tip percentage
-        EditNumber(
-            modifier = Modifier
-                .padding(bottom = 40.dp)
-                .align(alignment = Alignment.Start),
-            label = R.string.tip_percentage,
-            value = tipPercentInput,
-            onValueChange = { tipPercentInput = it },
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Number,
-                imeAction = ImeAction.Done
-            ),
-            leadingIcon = R.drawable.percent
-        )
-        RoundTheTipSwitch(roundUp, onRoundUpChange = { roundUp = it })
-        Text(
-            modifier = Modifier
-                .align(alignment = Alignment.Start)
-                .padding(bottom = 16.dp),
-            text = stringResource(R.string.tip_amount, tip),
-            style = MaterialTheme.typography.displaySmall,
-        )
-        Spacer(modifier = Modifier.height(150.dp))
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun EditNumber(
-    @StringRes label: Int,
-    @DrawableRes leadingIcon: Int,
-    modifier: Modifier = Modifier,
-    onValueChange: (String) -> Unit = {},
-    value: String,
-    keyboardOptions: KeyboardOptions
-) {
-    OutlinedTextField(
-        modifier = modifier.fillMaxWidth(),
-        singleLine = true,
-        leadingIcon = {
-            Icon(
-                painter = painterResource(leadingIcon),
-                contentDescription = null
-            )},
-        keyboardOptions = keyboardOptions,
-        label = { Text(stringResource(label)) },
-        value = value,
-        onValueChange = onValueChange,
-        // onValueChange = { input -> amountInput = input.filter { it.isDigit() }},
-        )
-}
-
-@Composable
-fun RoundTheTipSwitch(
-    roundUp: Boolean,
-    onRoundUpChange: (Boolean) -> Unit,
-    modifier: Modifier = Modifier
-)
-{
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .size(50.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ){
-        Text(text = stringResource(R.string.round_up_tip))
-        Switch(
-            modifier = Modifier
-                .fillMaxWidth()
-                .wrapContentWidth(Alignment.End),
-            checked = roundUp,
-            onCheckedChange = onRoundUpChange
-        )
-    }
-}
-
-private fun calculateTip(
-    amount: Double,
-    tipPercent: Double = 5.0,
-    roundUp: Boolean,
-): String {
-    var tip = tipPercent / 100 * amount
-    if(roundUp)
-        tip = kotlin.math.ceil(tip)
-    return NumberFormat.getCurrencyInstance().format(tip)
-}
 @Preview(showBackground = true)
 @Composable
 fun MatipPreview() {
     MaTIPv2Theme {
-        MatipLayout()
+        MainLayout()
     }
 }
